Check for circular dependencies during build, fix existing issues
