+++ lib/components/MenuItem/MenuItemLink.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
-import { Link } from '../';
-import { LinkProps } from '../Link/Link';
+import { Link, LinkProps } from '../Link/Link';
 import { useMenuItem } from './useMenuItem';
 import { MenuItemProps } from './MenuItem';
