+++ lib/components/MenuRenderer/MenuRendererItemContext.ts
@@ -1,0 +1,13 @@
+import { createContext } from 'react';
+import { Action } from './MenuRenderer.actions';
+
+interface MenuRendererItemContextValues {
+  isHighlighted: boolean;
+  index: number;
+  dispatch: (action: Action) => void;
+  focusTrigger: () => void;
+}
+
+export const MenuRendererItemContext = createContext<MenuRendererItemContextValues | null>(
+  null,
+);
