+++ lib/components/Tabs/TabListContext.ts
@@ -1,0 +1,8 @@
+import { createContext } from 'react';
+
+interface TabListContextValues {
+  tabListItemIndex: number;
+  scrollContainer: HTMLElement | null;
+}
+
+export const TabListContext = createContext<TabListContextValues | null>(null);
