+++ lib/components/Tabs/TabPanelsContext.ts
@@ -1,0 +1,10 @@
+import { createContext } from 'react';
+
+interface TabPanelsContextValues {
+  renderInactive: boolean;
+  panelIndex: number;
+}
+
+export const TabPanelsContext = createContext<TabPanelsContextValues | null>(
+  null,
+);
