+++ lib/components/useSpace/useSpace.ts
@@ -1,3 +1,3 @@
-import { useBraidTheme } from '../BraidProvider/BraidProvider';
+import { useBraidTheme } from '../BraidProvider/BraidThemeContext';
 
 export const useSpace = () => useBraidTheme().space;
