+++ lib/components/useThemeName/useThemeName.tsx
@@ -1,3 +1,3 @@
-import { useBraidTheme } from '../BraidProvider/BraidProvider';
+import { useBraidTheme } from '../BraidProvider/BraidThemeContext';
 
 export const useThemeName = () => useBraidTheme().name;
