+++ package.json
@@ -94,6 +94,7 @@
     "change-case": "^3.1.0",
     "cheerio": "^1.0.0-rc.3",
     "chromatic": "^5.1.0",
+    "circular-dependency-plugin": "^5.2.0",
     "concurrently": "^4.1.0",
     "copy-to-clipboard": "^3.2.0",
     "date-fns": "^2.15.0",
